ESQL: Deduplicate Parquet iterator decode logic
Extract shared decode helpers from ParquetColumnIterator and
OptimizedParquetColumnIterator into a new ParquetColumnDecoding
utility class. This consolidates list-column decoding, timestamp
conversion, UUID formatting, skip helpers, null bitmap conversion,
and NoOp converters so that bug fixes apply to both iterators
automatically instead of requiring parallel changes.

Removes ~530 lines of duplicated code, replacing with delegation
to the shared utility. The scalar ColumnReader decode path remains
in the baseline iterator since it is only used when record filters
are active.

Developed using AI-assisted tooling
