+++ docs/changelog/147584.yaml
@@ -1,0 +1,5 @@
+area: ES|QL
+issues: []
+pr: 147584
+summary: Deduplicate Parquet iterator decode logic
+type: enhancement
